requirements.txt
lxml>=4.9.0
html2text>=2020.1.16
PyYAML>=6.0
pathlib2>=2.3.7; python_version<"3.4"
click>=8.0.0
colorama>=0.4.4
tqdm>=4.64.0
pillow>=9.0.0
evernote-backup>=1.9.0
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0